Extract declaration-placement compiler fixtures
